SimpleScaling/requirements.txt
numpy~=2.2.0
Pillow~=11.0.0
